requirements.txt
flask==2.3.2
tensorflow==2.10.0
pillow==10.0.0
numpy==1.24.3